requirements.txt
python==3.10
numpy==1.23.5
pandas==1.5.3
matplotlib==3.7.0
pytorch==2.0.1+cu118